Sprints listos con tablas de sprint 5

+++ src/main/java/Controlador/Clientes.java
@@ -50,7 +50,7 @@ public class Clientes extends HttpServlet {
 			}
 		}
 		if (request.getParameter("consultar") != null) {			
-			int cedula = Integer.parseInt(request.getParameter("cedula"));						
+			long cedula = Integer.parseInt(request.getParameter("cedula"));						
 			clientesDto client = libdao.buscar_Cliente(cedula);
 			
 			if (client != null) {
